# packages/studio/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "InkOS Studio API",
    "version": "1.5.0",
    "description": "Auto-generated OpenAPI spec from 157 endpoints across 33 route modules."
  },
  "servers": [
    {
      "url": "http://localhost:4579",
      "description": "InkOS Studio API server"
    }
  ],
  "paths": {
...
    "/api/v1/books/{id}/import/chapters/plan": {
      "post": {
        "tags": [
          "import-foundation"
        ],
        "summary": "POST /api/v1/books/:id/import/chapters/plan",
        "operationId": "post__api_v1_books_id_import_chapters_plan",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/books/{id}/import/chapters/commit": {
      "post": {
        "tags": [
          "import-foundation"
        ],
        "summary": "POST /api/v1/books/:id/import/chapters/commit",
        "operationId": "post__api_v1_books_id_import_chapters_commit",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/books/{id}/import/chapters": {
      "post": {
        "tags": [
          "import-foundation"
        ],
        "summary": "POST /api/v1/books/:id/import/chapters",
        "operationId": "post__api_v1_books_id_import_chapters",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/books/{id}/import/canon": {
      "post": {
        "tags": [
          "import-foundation"
        ],
        "summary": "POST /api/v1/books/:id/import/canon",
        "operationId": "post__api_v1_books_id_import_canon",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/books/{id}/import/foundation/plan": {
      "post": {
        "tags": [
          "import-foundation"
        ],
        "summary": "POST /api/v1/books/:id/import/foundation/plan",
        "operationId": "post__api_v1_books_id_import_foundation_plan",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/books/{id}/import/foundation/commit": {
      "post": {
        "tags": [
          "import-foundation"
        ],
        "summary": "POST /api/v1/books/:id/import/foundation/commit",
        "operationId": "post__api_v1_books_id_import_foundation_commit",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/v1/books/{id}/revise/{chapter}": {
      "post": {
        "tags": [
          "revision-export"
        ],
        "summary": "POST /api/v1/books/:id/revise/:chapter",
        "operationId": "post__api_v1_books_id_revise_chapter",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/books/{id}/export": {
      "get": {
        "tags": [
          "revision-export"
        ],
        "summary": "GET /api/v1/books/:id/export",
        "operationId": "get__api_v1_books_id_export",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/books/{id}/export-save": {
      "post": {
        "tags": [
          "revision-export"
        ],
        "summary": "POST /api/v1/books/:id/export-save",
        "operationId": "post__api_v1_books_id_export_save",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/books/{id}/rewrite/{chapter}": {
      "post": {
        "tags": [
          "revision-export"
        ],
        "summary": "POST /api/v1/books/:id/rewrite/:chapter",
        "operationId": "post__api_v1_books_id_rewrite_chapter",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/books/{id}/resync/{chapter}": {
      "post": {
        "tags": [
          "revision-export"
        ],
        "summary": "POST /api/v1/books/:id/resync/:chapter",
        "operationId": "post__api_v1_books_id_resync_chapter",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}